Buggy multi-tap dance Ploopy keymap
This code has a bug. If you roll through any two of the tap dance keys, you'll get stuck in the held state for the second key. This will leave a layer or DRAG_SCROLL held on when no keys are pressed. I think this is a bug in QMK.